refactor: Criando pasta de rotas e alocando os arquivos de todolist, users, e tasks.
